requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.0

# Database
mysql-connector-python==8.2.0

# Document parsing libraries
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
lxml==4.9.3

# Vector database and embeddings (for future use)
pinecone-client==2.2.4
llama-index>=0.9.0
openai>=1.14.0

# Additional utilities
python-dateutil==2.8.2
pandas>=2.2.0
numpy>=1.26.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 